moved FormHeading component to component folder
